Alter break condition for adding catalysts
Previously the program would stop adding cataysts when the catalysts
already present were damaged for 5 generations consecutively. This
doesn't seem to make sense, especially when a transparent catalyst is
not the last catalyst to be activated.

Instead only quit adding catalysts when a "*" cell of one of the
existing catalysts dies. In particular, a transparent catalyst
disappearing does not prevent more catalysts being added later on.