Hack symmetry into ptbsearch
Transparent catalysts and minimum interaction generation have been
removed. Usgae is now:

ptb2 inputfile catalystsfile maxgen catalysts-1 mingen xcoord ycoord

The new parameters xcoord and ycoord define a 180 degree rotation such
that cell (0, 0) maps to cell (xcoord, ycoord).

At the start, and at every moment a catalyst is added, the pattern is
overlaid with the image of the pattern after the rotation, thus creating
symmetry at all times.

Because of this any catalysts that are 180 rotations of each other
should be avoided in the catalysts file as they will just be redundant.

Also 'survive' has been changed so that it only outputs patterns that
last for a long time without breaking and without becoming p2.